Add files via upload
This was a research project I began in early 2020 that culminated in the creation of an educational project that I have revised and modified for many different audiences depending on requirements and applicability to a security operation. This is one of many (early) incarnations of this project and I find myself reviving it on a yearly basis as its relevance has never failed to play a role in bringing staff up to speed on how C2 operates. No matter how far technology advances, Command and Control retains a death grip on offensive security practices. It is essential to actions on objectives for targets of significant intelligence and remains a primary choke-point for defenders to focus on.